Clean up 2015 day 2
Turns out that the slice module already has a split on a certain
character functionality, so I don't need to implement that. This might
be a case where it is simply easier to write the the code without using
iterators as heavly as I did, but it's fun trying to remove any mut in
the code.