2015 Day 6
So this one is the first one where things got verbose. It was mostly
because I wanted to make things safe. The task itself is quite boring,
but it was my first time playing with 2d arrays in rust.

This would maybe be better done with regex, but I managed without this
time.