2015 Day 7
This was a fun task, but it was a classic memorization task. It was good
to have one of these under my belt in rust, first time I've had to
explicitly set the lifetime of parameters and also the first time I've
seen the &mut prefix. I don't actually know what this means so I have to
google it.

Got part 2 pretty much for free from part 1.